UPSTREAM: tracing: Separate out initializing top level dir from instances
The top level trace array is treated a little different than the
instances, as it has to deal with more of the general tracing.
The tr->dir is the tracing directory, which is an immutable
dentry, where as the tr->dir of instances are the dentry that
was created, and can be destroyed later. These should have different
functions accessing them.

As only tracing_init_dentry() deals with the top level array, fold
the code for it into that function, and remove the trace_init_dentry_tr()
that was also used by the instances to get their directory dentry.

Add a tracing_get_dentry() to just get the tracing dir entry for
instances as well as the top level array.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>

Bug: 31856701
Change-Id: I8aa90d48d9e4a23ad9972b73c491e9d5f3d83f3c
(cherry picked from commit 7eeafbcab47fe9860e5106286db83d60e3f35644)
Signed-off-by: Paul Lawrence <paullawrence@google.com>
Signed-off-by: Subhaam <derksoul98@gmail.com>