UPSTREAM: procfs: fixes pthread cross-thread naming if !PR_DUMPABLE
The PR_DUMPABLE flag causes the pid related paths of the
proc file system to be owned by ROOT. The implementation
of pthread_set/getname_np however needs access to
/proc/<pid>/task/<tid>/comm.
If PR_DUMPABLE is false this implementation is locked out.

This patch installs a special permission function for
the file "comm" that grants read and write access to
all threads of the same group regardless of the ownership
of the inode. For all other threads the function falls back
to the generic inode permission check.

Bug: 30067360
Signed-off-by: Janis Danisevskis <jdanis@google.com>
Change-Id: I4175a341c5a598b44da9723a0e7af3c324ac9016
Signed-off-by: Subhaam <derksoul98@gmail.com>